# refactor(filetype): extract some functions, use more cache