There are three parts in this refactor/optimization:
- Separate single "sorted patterns" list into two: "sorted patterns with non-negative priority" and "sorted patterns with negative priority". It allows extracting logic of matching sorted patterns into a separate function. As a result, main `vim.filetype.match()` code is cleaner and pattern matching is easier to manage (hopefully, will be important in future PRs).
- Instead of caching only whether pattern has environment variable (during "match" time), cache other data about it (during "add" time): full pattern (with '^' / '$' at start / end) and whether it contains slash. use them in equivalent manner as before.
- Extract "expand environment variables" in a separate function. I tried simplifying it by simply throwing error if there is no envvar defined (and later call it with `pcall()`), but this seemed to be slower than current approach.
- Convert patterns to be explicit (without implicit anchoring). This proved to be slightly faster (probably due to less backtracking during match) and allows to not compute and cache `fullpat`.

------

Each step also seems to lead to small but consistent performance improvements. I tested (with debug build) for three cases of 'filename' on `main` and latest commit of this PR:

Filename | `master` | This PR |
--------------|-----------|--------------|
init.lua | 0.136ms  | 0.111ms |
no-match | 0.163ms | 0.132ms |
/home/user/.config/hello/world/no-match | 0.240ms | 0.210ms |

------

This is all done in (hopefully) a preparation for #29660.