@echasnovski 2024-07-12
@gpanders, I've force pushed several changes that were suggested in #29660. Would you mind re-review, please?

List of changes:
- Caching pattern data is done during "add" time. [Relevant suggestion](https://github.com/neovim/neovim/pull/29660#discussion_r1675514653). Also renamed several local variables to be more explicit.
- All patterns in `pattern` table are now explicit (without implicit anchoring). I planned to suggest it later (after "fast pattern" PR), but here is a [relevant suggestion](https://github.com/neovim/neovim/pull/29660#discussion_r1675510985).

Also rebased on current `master` and updated first comment with more up to date info. Other things are untouched.